Convert pulumi-onelogin to a patched provider
Before this PR, the provider doesn't actually take a dependency on the upstream
provider. This makes it hard to figure out the current version from examining the repo
contents and impossible for automated upgrades to work (there is nothing to upgrade).

After this PR, we can perform upgrades as with any normal patched provider, and automatic
upgrade tooling should work as expected.